+++ template/ios/HelloWorld/AppDelegate.swift
@@ -3,7 +3,6 @@ import React
 import React_RCTAppDelegate
 import ReactAppDependencyProvider
 import GoogleMaps
-import Firebase
 import RNBootSplash
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -17,7 +16,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil
   ) -> Bool {
     // Initialize Google Maps
-    FirebaseApp.configure()
     GMSServices.provideAPIKey("<GOOGLE_MAPS_API_KEY>")
     
     let delegate = ReactNativeDelegate()